backend/requirements.txt
# AI-Based Lecture Summarizer - Backend Dependencies
# Python 3.8+ required

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# OpenAI Whisper for Speech Recognition
openai-whisper==20231117

# Audio Processing
numpy==1.24.3
torch==2.1.0

# NLP & Transformers (Module 2)
transformers==4.35.0
sentencepiece==0.1.99
protobuf==4.25.1

# BART Summarization (Module 2)
# Note: BART uses the same transformers library

# Text Processing
nltk==3.8.1
scikit-learn==1.3.2
rake-nltk==1.0.6

# Utilities
python-dotenv==1.0.0

# Optional: For better tokenization
tokenizers==0.15.0
